fixes based on comments
 * b felony test cases back in the shared time analyzer test file
 * rewrite the 20-year limit logic for readability
 * 'over 20 years ago' = 20 years + 1 day, not 21 years

also:
 * subsequent arrests cause time eligibility, not type eligibility of a B felony
