+++ src/backend/expungeservice/expunger/analyzers/time_analyzer.py
@@ -6,7 +6,6 @@ from expungeservice.models.expungement_result import TypeEligibility, Expungemen
 
 class TimeAnalyzer:
 
-    TWENTY_YEARS_AGO = date.today() + relativedelta(years=-20)
     TWENTY_YEARS = 20
     TEN_YEARS = 10
     THREE_YEARS = 3
@@ -77,20 +76,22 @@ class TimeAnalyzer:
     def _evaluate_class_b_felonies(expunger):
         for class_b_felony in expunger.class_b_felonies:
             # If the class B felony is acquitted, then we have already handled its time eligibility.
-            # If it's convicted, this restriction overrides any of those rules.
-            if not class_b_felony.acquitted():
+            # If it's convicted, it is subject to these additional rules.
+            # If its disposition is unknown, we don't apply any time analysis.
+            if class_b_felony.convicted():
                 if TimeAnalyzer._calculate_has_subsequent_charge(class_b_felony, expunger.charges):
-                    # TODO: Express this with TimeEligibility with no eligibility date.
-                    type_eligibility = TypeEligibility(
-                        EligibilityStatus.INELIGIBLE,
-                        reason='137.225(5)(a)(A)(ii) - Class B felony can have no subsequent arrests or convictions')
-                    class_b_felony.expungement_result = ExpungementResult(type_eligibility=type_eligibility, time_eligibility=None)
-
+                    # type_eligibility = TypeEligibility(EligibilityStatus.INELIGIBLE, reason = '137.225(5)(a)(A)(ii) - Class B felony can have no subsequent arrests or convictions')
+                    # class_b_felony.expungement_result = ExpungementResult(type_eligibility=type_eligibility, time_eligibility=None)
+                    # TODO:
+                    # Apply time ineligibility here instead of type eligibility.
+                    # The current logic here is just to make the frontend display correctly until we fix it there.
+                    # The new logic here will be:
+                    class_b_felony.set_time_ineligible('137.225(5)(a)(A)(ii) - Class B felony can have no subsequent arrests or convictions', None)
                 else:
-                    if class_b_felony.disposition.date > TimeAnalyzer.TWENTY_YEARS_AGO:
-                        class_b_felony.set_time_ineligible('137.225(5)(a)(A)(i) - Twenty years from class B felony conviction',
-                                                       class_b_felony.disposition.date + relativedelta(
-                                                           years=TimeAnalyzer.TWENTY_YEARS))
+                    eligibility_date = class_b_felony.disposition.date + relativedelta(years=+TimeAnalyzer.TWENTY_YEARS)
+                    if eligibility_date > date.today():
+                        class_b_felony.set_time_ineligible('137.225(5)(a)(A)(i) - Twenty years from class B felony conviction', eligibility_date)
+                    # else do nothing, because the charge has already been set as time eligible by the TimeAnalyzer.
 
     @staticmethod
     def _mark_as_time_ineligible(charges, reason, eligibility_date):
